move pattern render to comp
